refactor CustomRuleFormModal
